# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EventMaster API
  description: |
    活動報名與驗票系統 Backend API

    EventMaster API provides comprehensive event registration and check-in management
    with QR code-based ticket verification. Built with FastAPI and designed for
    small to medium-scale events (<200 concurrent users).

    ## Features
    - JWT-based authentication with secure token management
    - Role-based access control (member, organizer, admin)
    - Event creation and management
    - Automated QR code ticket generation
    - Real-time check-in verification
    - Walk-in registration support
    - Paginated list endpoints for optimal performance

    ## Authentication
    使用 JWT Bearer Token 進行驗證。在 Authorization header 中傳入 token:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Role-Based Access Control (RBAC)
    - **member**: 一般使用者，可報名活動、查看自己的報名與 QR Code
    - **organizer**: 主辦/櫃台人員，可查看報名名單並進行驗票
    - **admin**: 系統管理者，可建立活動與指派角色

    ## Rate Limiting
    - 100 requests per minute per authenticated user
    - 1000 requests per hour per IP address

    ## Support
    For API support, please contact support@eventmaster.com or visit our
    [support portal](https://eventmaster.com/support).

  version: 1.0.0

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

  contact:
    name: EventMaster Team
    email: support@eventmaster.com
    url: https://eventmaster.com/support

  termsOfService: https://eventmaster.com/terms
servers:
...
security:
  - BearerAuth: []
paths:
...
  /events:
    get:
      operationId: listEvents
      tags:
        - Events
      summary: 取得活動列表
      description: |
        取得活動列表 (依角色顯示)
        - 未登入/Member: 只顯示 PUBLISHED
        - Organizer: PUBLISHED + 自己建立的活動 (所有狀態)
        - Admin: 所有活動 (所有狀態)
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 成功取得活動列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - limit
                  - offset
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
                    description: 總筆數 (所有頁面的總數量)
                    example: 42
                  limit:
                    type: integer
                    description: 每頁最大筆數
                    example: 20
                  offset:
                    type: integer
                    description: 跳過的筆數
                    example: 0

    post:
      operationId: createEvent
      tags:
        - Events
      summary: 建立新活動
      description: |
        建立新活動 (需要 Admin 或 Organizer 權限)
        - Organizer 建立 → status = PENDING
        - Admin 建立 → status = PUBLISHED
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
            examples:
              quarterlyMeeting:
                summary: Quarterly all-hands meeting
                value:
                  title: "Q1 All-Hands Meeting"
                  description: "Company wide updates and quarterly goals review."
                  startAt: "2024-03-15T10:00:00Z"
                  endAt: "2024-03-15T12:00:00Z"
                  location: "Main Auditorium"
                  capacity: 200
              techWorkshop:
                summary: Technical workshop
                value:
                  title: "Advanced FastAPI Workshop"
                  description: "Hands-on workshop covering FastAPI best practices, async patterns, and testing strategies."
                  startAt: "2024-04-20T14:00:00Z"
                  endAt: "2024-04-20T17:00:00Z"
                  location: "Training Room B"
                  capacity: 30
      responses:
        '201':
          description: 活動建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
              examples:
                eventCreated:
                  summary: Event created successfully
                  value:
                    id: "a1b2c3d4-e5f6-4a5b-8c9d-1e2f3a4b5c6d"
                    organizerId: "550e8400-e29b-41d4-a716-446655440002"
                    title: "Q1 All-Hands Meeting"
                    description: "Company wide updates and quarterly goals review."
                    startAt: "2024-03-15T10:00:00Z"
                    endAt: "2024-03-15T12:00:00Z"
                    location: "Main Auditorium"
                    capacity: 200
                    registeredCount: 0
                    status: "PENDING"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /events/pending:
    get:
      operationId: getPendingEvents
      tags:
        - Events
      summary: 取得待審核活動
      description: 取得所有 PENDING 狀態的活動 (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 成功取得待審核活動列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - limit
                  - offset
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
                    description: 待審核活動總數
                    example: 3
                  limit:
                    type: integer
                    example: 20
                  offset:
                    type: integer
                    example: 0
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /events/managed:
    get:
      operationId: getManagedEvents
      tags:
        - Events
      summary: 取得自己管理的活動
      description: 取得當前使用者管理的活動 (Organizer 取得自己建立的活動，Admin 取得所有活動)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: 成功取得管理的活動列表
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - limit
                  - offset
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
                    description: 總筆數
                    example: 5
                  limit:
                    type: integer
                    description: 每頁最大筆數
                    example: 20
                  offset:
                    type: integer
                    description: 跳過的筆數
                    example: 0
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /events/{eventId}:
    get:
      operationId: getEvent
      tags:
        - Events
      summary: 取得活動詳細資訊
      description: 根據活動 ID 取得活動詳細資訊
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '200':
          description: 成功取得活動資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      operationId: updateEvent
      tags:
        - Events
      summary: 更新活動資訊
      description: 更新活動資訊 (需要是活動的 Organizer 或 Admin)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdateRequest'
      responses:
        '200':
          description: 活動更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      operationId: deleteEvent
      tags:
        - Events
      summary: 刪除活動
      description: 刪除活動 (需要是活動的 Organizer 或 Admin)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '204':
          description: 活動刪除成功
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{eventId}/approve:
    patch:
      operationId: approveEvent
      tags:
        - Events
      summary: 核准活動
      description: 將活動狀態從 PENDING 變更為 PUBLISHED (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '200':
          description: 活動核准成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: 活動不是 PENDING 狀態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /events/{eventId}/reject:
    patch:
      operationId: rejectEvent
      tags:
        - Events
      summary: 駁回活動
      description: 將活動狀態從 PENDING 變更為 REJECTED (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventIdParam'
      responses:
        '200':
          description: 活動駁回成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: 活動不是 PENDING 狀態
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==================== Registrations ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        使用 JWT Bearer Token 進行驗證
        格式: `Authorization: Bearer <token>`
  parameters:
    EventIdParam:
      name: eventId
      in: path
      required: true
      description: 活動 ID
      schema:
        type: string
        format: uuid
        example: "a1b2c3d4-e5f6-4a5b-8c9d-1e2f3a4b5c6d"
    LimitParam:
      name: limit
      in: query
      description: 每頁最大筆數
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      example: 20
    OffsetParam:
      name: offset
      in: query
      description: 跳過筆數 (用於分頁)
      schema:
        type: integer
        default: 0
        minimum: 0
      example: 0
  schemas:
...
    EventStatus:
      type: string
      enum:
        - PENDING
        - PUBLISHED
        - REJECTED
      description: |
        活動審核狀態:
        - `PENDING`: 待審核
        - `PUBLISHED`: 已發布
        - `REJECTED`: 已駁回
    Event:
      type: object
      required:
        - id
        - organizerId
        - title
        - description
        - startAt
        - endAt
        - location
        - capacity
        - registeredCount
        - status
      properties:
        id:
          type: string
          format: uuid
          description: 活動 ID
          example: "a1b2c3d4-e5f6-4a5b-8c9d-1e2f3a4b5c6d"
        organizerId:
          type: string
          format: uuid
          description: 主辦方 User ID
          example: "550e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          description: 活動標題
          example: Q1 All-Hands Meeting
        description:
          type: string
          description: 活動描述
          example: Company wide updates and quarterly goals review.
        startAt:
          type: string
          format: date-time
          description: 活動開始時間 (ISO 8601)
          example: "2023-11-15T10:00:00Z"
        endAt:
          type: string
          format: date-time
          description: 活動結束時間 (ISO 8601)
          example: "2023-11-15T12:00:00Z"
        location:
          type: string
          description: 活動地點
          example: Main Auditorium
        capacity:
          type: integer
          description: 活動容納人數
          minimum: 1
          example: 200
        registeredCount:
          type: integer
          description: 已報名人數
          minimum: 0
          example: 45
        status:
          $ref: '#/components/schemas/EventStatus'
    EventCreateRequest:
      type: object
      required:
        - title
        - description
        - startAt
        - endAt
        - location
        - capacity
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Q1 All-Hands Meeting
        description:
          type: string
          minLength: 1
          maxLength: 2000
          example: Company wide updates and quarterly goals review.
        startAt:
          type: string
          format: date-time
          example: "2023-11-15T10:00:00Z"
        endAt:
          type: string
          format: date-time
          example: "2023-11-15T12:00:00Z"
        location:
          type: string
          minLength: 1
          maxLength: 200
          example: Main Auditorium
        capacity:
          type: integer
          minimum: 1
          maximum: 10000
          example: 200
    EventUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 1
          maxLength: 2000
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        location:
          type: string
          minLength: 1
          maxLength: 200
        capacity:
          type: integer
          minimum: 1
          maximum: 10000

    # ========== Registration Related ==========
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: 錯誤類型
          example: ValidationError
        message:
          type: string
          description: 錯誤訊息
          example: Invalid input data
        details:
          type: object
          description: 詳細錯誤資訊 (選填)
          additionalProperties: true
  responses:
    UnauthorizedError:
      description: 未認證 - Token 無效或未提供
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing authentication token
              value:
                error: Unauthorized
                message: Invalid or missing authentication token
            expiredToken:
              summary: Expired token
              value:
                error: Unauthorized
                message: Token has expired
    ForbiddenError:
      description: 權限不足 - 使用者沒有執行此操作的權限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Insufficient permissions
              value:
                error: Forbidden
                message: You do not have permission to perform this action
            wrongRole:
              summary: Wrong user role
              value:
                error: Forbidden
                message: This action requires organizer or admin role
    NotFoundError:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                error: NotFound
                message: Resource not found
    ValidationError:
      description: 輸入驗證錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationFailure:
              summary: Input validation failure
              value:
                error: ValidationError
                message: Invalid input data
                details:
                  email: Invalid email format
                  capacity: Must be greater than 0